pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-resume_env"
version = "0.1.0"
description = "Resume Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    # OpenAI-compatible client for LLM inference
    "openai>=1.0.0",
    # Load .env files for local development
    "python-dotenv>=1.0.0",
    # HTTP requests (used by validate_submission.py)
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m resume_env.server.app
server = "resume_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["resume_env", "resume_env.server"]
package-dir = { "resume_env" = ".", "resume_env.server" = "server" }